tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/child_navigates_parent_submit.html

<!doctype html>
<title>Child document navigating parent via submit </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test();
var events = [];
onmessage = t.step_func(function(e) {
  events.push(e.data);
  if (events.length == 3) {
    assert_array_equals(events, ["initial", "inner", "destination"]);

    t.done();
  }
});
</script>
<iframe src="child_navigates_parent_submit-1.html"></iframe>
